make trees more balanced
this wasn't in the description of the huffman algorithm I found, but
adding 1 'weight' per tree depth should force the trees to be more
balanced (since all the smallest trees will get paired up first, rather
than in the random order they come out of the priority queue). This
shortens the encodings of things CONSIDERABLY (by a factor of like 2-3.)
However, it seems to have broken the "shortest thing encodes to 1 bit"
rule. So I'll have to look at  that.